Bug Fixed

+++ Core/src/main/java/uz/muhammadyusuf/kurbonov/core/repos/Repository.kt
@@ -13,5 +13,5 @@ internal interface Repository {
 
     suspend fun getItem(id: Int): AccountingItem
 
-    suspend fun calculateSum(): Int
+    suspend fun calculateSum(): Double
 }
